separate composable content from TrustbadgeView
