+++ library/src/main/java/com/etrusted/android/trustbadge/library/ui/badge/TrustbadgeView.kt
@@ -25,33 +25,12 @@
 
 package com.etrusted.android.trustbadge.library.ui.badge
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.animateContentSize
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.defaultMinSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.ElevatedButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.etrusted.android.trustbadge.library.common.internal.ExcludeFromJacocoGeneratedReport
-import com.etrusted.android.trustbadge.library.common.internal.TestTags
 import com.etrusted.android.trustbadge.library.ui.badge.TrustbadgeContext.ShopGrade
 import com.etrusted.android.trustbadge.library.ui.theme.TrustbadgeTheme
-import com.etrusted.android.trustbadge.library.ui.theme.TsBadgeBg
-import com.etrusted.android.trustbadge.library.ui.theme.TsNeutralsGrey50
-import kotlinx.coroutines.delay
 
 @ExcludeFromJacocoGeneratedReport
 @Composable
@@ -64,81 +43,6 @@ fun Trustbadge(
 ) {
     TrustbadgeContent(modifier = modifier, state = state, badgeContext = badgeContext,
         tsid = tsid, channelId = channelId)
-}
-
-/**
- * Trustbadge content allowing usage of custom viewModel in tests
- */
-@Composable
-internal fun TrustbadgeContent(
-    modifier: Modifier,
-    viewModel: ITrustbadgeViewModel = viewModel<TrustbadgeViewModel>(),
-    state: TrustbadgeState,
-    badgeContext: TrustbadgeContext,
-    tsid: String,
-    channelId: String
-) {
-    val trustbadgeData by viewModel.trustbadgeData.collectAsState()
-    val guarantee by viewModel.guarantee.collectAsState()
-
-    AnimatedVisibility(
-        modifier = modifier.testTag(TestTags.Trustbadge.raw),
-        visible = state.currentState != TrustbadgeStateValue.INVISIBLE
-    ) {
-
-        ElevatedButton(
-            modifier = Modifier
-                .animateContentSize()
-                .padding(horizontal = 24.dp, vertical = 24.dp)
-                .height(66.dp)
-                .defaultMinSize(minWidth = 66.dp),
-            contentPadding = PaddingValues(start = 66.dp),
-            onClick = {},
-            border = BorderStroke(1.dp, MaterialTheme.colorScheme.TsNeutralsGrey50),
-            colors = ButtonDefaults.elevatedButtonColors(
-                containerColor = MaterialTheme.colorScheme.TsBadgeBg
-            )
-        ) {
-            ExpandedView(
-                modifier = Modifier.align(Alignment.CenterVertically),
-                state = state, badgeContext = badgeContext,
-                rating = trustbadgeData?.shop?.rating,
-                guaranteeAmount = guarantee?.maxProtectionAmount ?: "0"
-            )
-        }
-
-        ElevatedButton(
-            modifier = Modifier
-                .padding(horizontal = 24.dp, vertical = 24.dp),
-            contentPadding = PaddingValues(0.dp),
-            onClick = {},
-            colors = ButtonDefaults.elevatedButtonColors(
-                containerColor = MaterialTheme.colorScheme.TsBadgeBg
-            )
-        ) {
-
-            RoundedView(
-                modifier = Modifier.testTag(TestTags.TrustbadgeDefault.raw),
-                state = state, badgeContext = badgeContext
-            )
-        }
-    }
-
-    LaunchedEffect(null) {
-        viewModel.fetchTrustbadgeData(tsid, channelId)
-        viewModel.fetchGuarantee(tsid, channelId)
-    }
-
-    LaunchedEffect(null) {
-        // automatically show the expanded state only if the context is not set to TRUSTMARK
-        // The TRUSTMARK state only shows the badge in circle form
-        if (badgeContext.isExpandable) {
-            delay(1000)
-            state.expand()
-            delay(3000)
-            state.retract()
-        }
-    }
 }
 
 @ExcludeFromJacocoGeneratedReport
